requirements.txt
Scrapy==2.11.2
python-dotenv==1.0.1